Fix 9 non-matching functions: narrow return type u64 → u32 for mov w0,wzr codegen
These vtable dispatch wrappers return 0 but were typed as u64, causing
the compiler to emit `mov x0, xzr` (64-bit) instead of `mov w0, wzr` (32-bit).
Changing return type to u32 produces matching codegen.

Functions fixed:
- FUN_710335d314, FUN_7103366688 (fun_batch_d_002)
- FUN_710335a5fc (fun_batch_d_003)
- FUN_710335c924 (fun_batch_d_005)
- FUN_710335c478, FUN_7103360268, FUN_7103365d08 (fun_batch_d_006)
- FUN_710373e054 (fun_batch_d_014, via auto_fix)
- FUN_71033625b8 (fun_batch_d_016)

+++ src/app/fun_batch_d_006.cpp
@@ -141,8 +141,8 @@ u64 *FUN_71037a02a0(void)
     return (u64*)DAT_710593ac48;
 }
 
-// 0x710335c478 -- vtable call at 0x268 (no args), return 0 (300 bytes)
-u64 FUN_710335c478(s64 *param_1)
+// 0x710335c478 -- vtable call at 0x268 (no args), return 0 (32 bytes)
+u32 FUN_710335c478(s64 *param_1)
 {
     (*(void(*)())(*(s64 *)(*param_1 + 0x268)))();
     return 0;
@@ -155,15 +155,15 @@ u64 FUN_71031d04b0(s64 *param_1)
     return 1;
 }
 
-// 0x7103360268 -- vtable call at 8 (no args), return 0 (352 bytes)
-u64 FUN_7103360268(s64 *param_1)
+// 0x7103360268 -- vtable call at 8 (no args), return 0 (32 bytes)
+u32 FUN_7103360268(s64 *param_1)
 {
     (*(void(*)())(*(s64 *)(*param_1 + 8)))();
     return 0;
 }
 
-// 0x7103365d08 -- vtable call at 8 (no args), return 0 (352 bytes)
-u64 FUN_7103365d08(s64 *param_1)
+// 0x7103365d08 -- vtable call at 8 (no args), return 0 (32 bytes)
+u32 FUN_7103365d08(s64 *param_1)
 {
     (*(void(*)())(*(s64 *)(*param_1 + 8)))();
     return 0;
